requirements.txt
librosa>=0.10.2
numpy>=1.26.0
scipy>=1.11.0
plotly>=5.20.0
dtw-python>=1.5.1
soundfile>=0.12.1
